Y-B synchronized

+++ services/abilitymgr/src/ability_record.cpp
@@ -271,7 +271,8 @@ bool AbilityRecord::CanRestartResident()
             abilityMgr->GetRestartIntervalTime(restartIntervalTime);
         }
         HILOG_DEBUG("restartTime: %{public}lld, now: %{public}lld, intervalTine:%{public}d",
-            restartTime_, AbilityUtil::SystemTimeMillis(), restartIntervalTime);
+            static_cast<unsigned long long>(restartTime_),
+            static_cast<unsigned long long>(AbilityUtil::SystemTimeMillis()), restartIntervalTime);
         if ((AbilityUtil::SystemTimeMillis() - restartTime_) < restartIntervalTime) {
             HILOG_ERROR("Resident restart is out of max count");
             return false;
